belong for folders check belonging before file handling

+++ SeedDMS_Core/Core/inc.ClassGitExport.php
@@ -135,6 +135,11 @@ class SeedDMS_Core_Git_Export { /* {{{ */
 	  if ($document->getAttributeValue("ignoreInGit") == "true")
 	    return false;
 	  $curr = $document->getFolder();
+	  return $this->belongsFolderToRepository($curr);
+	}
+	
+	function belongsFolderToRepository($folder){
+	  $curr = $folder;
 	  while (true){
 	    if (!$curr)
 	      break;
@@ -153,6 +158,8 @@ class SeedDMS_Core_Git_Export { /* {{{ */
 	}
 	
 	function addDocumentContent($document){//todo: alter content bleibt beibehalten, wenn sich Dateityp ändert
+		if (!$this->belongsFileToRepository($document))
+		  return false;
 		$destinationPath = $this->DocumentGetGitPath($document);
 		$this->forceDirectories($destinationPath);
 		$this->log("Adding Document ".$document->getName());
@@ -171,6 +178,8 @@ class SeedDMS_Core_Git_Export { /* {{{ */
 	}
 	
 	function renameDocument($document, $oldname, $newname){
+		if (!$this->belongsFileToRepository($document))
+		  return false;
 		//$this->log($this->getGitStatus());
 		$documentPath = $this->DocumentGetGitPath($document);
 		$oldGitFile = $documentPath.'/'.$oldname.$document->getLatestContent()->getFileType();
@@ -192,6 +201,8 @@ class SeedDMS_Core_Git_Export { /* {{{ */
 	}
 	
 	function renameFolder($folder, $oldname, $newname){
+		if (!$this->belongsFolderToRepository($folder))
+		  return false;
 		$newpath = $this->FolderGetRelativePath($folder);//ist das inklusive ordnernamen selbst?
 		$oldpath = dirname($newpath)."/".$newname;
 		$this->log("Renaming Folder ".$oldpath." to ".$newpath);
@@ -213,6 +224,8 @@ class SeedDMS_Core_Git_Export { /* {{{ */
 	}
 	
 	function removeDocument($document, $latestContent){
+		if (!$this->belongsFileToRepository($document))
+		  return false;
 		if(unlink($this->DocumentGetGitFullPath($document,$latestContent))){
 			$this->gitRemove($this->DocumentGetGitFullPath($document,$latestContent));
 			$this->_gitCommitMessage .= "removed File ".$document->getName()."\r\n";
@@ -225,6 +238,8 @@ class SeedDMS_Core_Git_Export { /* {{{ */
 	}
 	
 	function removeFolder($folder){
+		if (!$this->belongsFolderToRepository($folder))
+		  return false;
 		if(unlink($this->FolderGetGitFullPath($folder))){
 			$this->gitRemove($this->FolderGetGitFullPath($folder), true);
 			$this->_gitCommitMessage .= "removed Folder".$folder->getName()."\r\n";
@@ -242,6 +257,8 @@ class SeedDMS_Core_Git_Export { /* {{{ */
 	 * @var integer newparent id of new parent folder
 	*/
 	function moveDocument($document, $oldparent, $newparent){
+		if (!$this->belongsFileToRepository($document))
+		  return false;
 		$oldFolder = $this->_dms->getFolder($oldparent);
 		$newFolder = $this->_dms->getFolder($newparent);
 		$destinationPath = $this->FolderGetGitFullPath($newFolder);
@@ -268,6 +285,9 @@ class SeedDMS_Core_Git_Export { /* {{{ */
 	
 	function moveFolder($folder, $oldparent, $newparent){
 		$this->log("Kann Ordner noch nicht bewegen", PEAR_LOG_WARN);
+		return false;
+		if (!$this->belongsFolderToRepository($folder))
+		  return false;
 		//rename();
 		//xml setzen auch für alle child elemente
 	}
